+++ CHANGELOG.md
@@ -4,6 +4,9 @@
 ### Added
 - `Node#new` method, creates a new instance of the same type using the same configuration (such as `uid`).
 
+### Changed
+- `Node#read` has been renamed to `Node#value`.
+
 ### Removed
 - `Node#clone` has been replaced in favor of `Node#new`, which does not copy properties.
 - `Graph#add` has been removed in favor of `Graph#merge`.